+++ src/components/Admin/PropertyCard.jsx
@@ -9,6 +9,7 @@ const PropertyCard = ({ property }) => {
   };
 
   return (
+	
     <div className="bg-white p-4 rounded-lg shadow-md hover:shadow-lg transition-shadow">
       {property.image && (
         <img
@@ -26,7 +27,12 @@ const PropertyCard = ({ property }) => {
       >
         View Details
       </button>
+	 
+	 
     </div>
+
+
+
   );
 };
