+++ src/components/Contact.jsx
@@ -23,7 +23,7 @@ const Contact = () => {
   const handleSubmit = async (e) => {
     e.preventDefault();
     try {
-      const response = await axios.post("http://localhost:3000/api/create-message", formData);
+      const response = await axios.post("https://easy-renting-bn.onrender.com/api/create-message", formData);
       setResponseMessage(response.data.message);
       setErrorMessage(""); 
       setFormData({ username: "", email: "", message: "" });
