+++ src/pages/RequestedProperty.jsx
@@ -48,7 +48,7 @@ export default function RequestedProperties() {
 
       <button onClick={()=>{
 		navigate('/create-request-property')
-	  }} className='rounded-lg text-xl text-white flex items-center bg-green-500 hover:bg-green-600 py-2 px-4 mt-10 lg:w-full min-w-[300px] mx-auto'>Supply Property</button>
+	  }} className='rounded-lg text-xl text-white  text-center bg-green-500 hover:bg-green-600 py-2 px-4 mt-10 lg:w-full max-w-[300px] mx-auto'>Request your Property</button>
     </div>
   );
 };
